feat(testing): Split out converter to avoid dependency on dd-trace-api and internal-api
