+++ dd-trace-core/build.gradle
@@ -119,7 +119,8 @@ dependencies {
   testImplementation group: 'com.amazonaws', name: 'aws-lambda-java-events', version:'3.11.0'
   testImplementation group: 'com.google.protobuf', name: 'protobuf-java', version: '3.14.0'
   testImplementation libs.testcontainers
-  testImplementation project(':utils:junit-utils')
+  testImplementation project(':utils:test-junit-utils')
+  testImplementation project(':utils:test-junit-converter-utils')
 
   traceAgentTestImplementation libs.testcontainers
 }
