+++ dd-trace-ot/build.gradle.kts
@@ -64,11 +64,14 @@ dependencies {
   implementation(project(":dd-trace-ot:correlation-id-injection"))
 
   testImplementation(project(":dd-java-agent:testing"))
-  testImplementation(project(":utils:junit-utils"))
+  testImplementation(project(":utils:test-junit-utils"))
+  testImplementation(project(":utils:test-junit-converter-utils"))
   testImplementation(libs.bundles.mockito)
 
-  add("ot31CompatibilityTestImplementation", project(":utils:junit-utils"))
-  add("ot33CompatibilityTestImplementation", project(":utils:junit-utils"))
+  add("ot31CompatibilityTestImplementation", project(":utils:test-junit-utils"))
+  add("ot31CompatibilityTestImplementation", project(":utils:test-junit-converter-utils"))
+  add("ot33CompatibilityTestImplementation", project(":utils:test-junit-utils"))
+  add("ot33CompatibilityTestImplementation", project(":utils:test-junit-converter-utils"))
 
   // Kotlin accessors not generated if not coming from plugin
   add("ot33CompatibilityTestImplementation", "io.opentracing:opentracing-api") {
