+++ settings.gradle.kts
@@ -170,7 +170,8 @@ include(
   ":dd-java-agent:testing",
   ":utils:config-utils",
   ":utils:container-utils",
-  ":utils:junit-utils",
+  ":utils:test-junit-utils",
+  ":utils:test-junit-converter-utils",
   ":utils:filesystem-utils",
   ":utils:flare-utils",
   ":utils:logging-utils",
