+++ utils/test-junit-converter-utils/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+  `java-library`
+}
+
+apply(from = "$rootDir/gradle/java.gradle")
+
+dependencies {
+  implementation(project(":dd-trace-api"))
+  implementation(project(":internal-api"))
+
+  compileOnly(libs.junit.jupiter)
+  compileOnly(libs.tabletest)
+}
