+++ utils/test-junit-utils/build.gradle.kts
@@ -8,9 +8,5 @@ dependencies {
   api(libs.forbiddenapis)
   api(project(":components:environment"))
 
-  implementation(project(":dd-trace-api"))
-  implementation(project(":internal-api"))
-
   compileOnly(libs.junit.jupiter)
-  compileOnly(libs.tabletest)
 }
